Parser/requirements.txt
# requirements.txt

# Core
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# Graph database
neo4j==5.15.0

# Telegram
telethon==1.34.0
cryptg==0.4.0  # Speed up Telethon

# Web framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
httpx==0.26.0

# Message queue
aio-pika==9.4.0
pika==1.3.2

# Cache
redis==5.0.1
aioredis==2.0.1

# NLP & NER
natasha==1.6.0
pymorphy3==1.2.1
pymorphy3-dicts-ru==2.4.417150.4580142

# Text processing
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
transliterate==1.10.2

# Images
pillow==10.1.0
imagehash==4.3.1  # For image deduplication

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0
python-json-logger==2.0.7

# Utilities
pyyaml==6.0.1
python-dateutil==2.8.2
aiofiles==23.2.1
tenacity==8.2.3  # For retries

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0
mypy==1.7.1
ipython==8.18.1